Fix #67: Download is rate-limited by GitHub
Store release JSON data in repository, and use it to download rustywind.

We still use the GitHub API when downloading rustywind; if it fails, we
try downloading from the "browser download url" which is not on GitHub's
api subdomain. Whether this fall-back will actually help depends on how
GitHub's rate-limiting works.